Keep SniffAir from crashing during EAP import
While importing a pcap file, SniffAir would crash when trying to parse a
Request, Identity packet. What we are looking for are Response, Identity
packets. Both packets contained the Type: Identity (1); however, The
Code: for Request is 1 and 2 for Response. This commit will look for the
Code: 2.

I've also added an exception catch to dump the packet that causes the
error and exits the import.